+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @kaitox/cli
 
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @kaitox/x-article@0.6.0
+
 ## 0.4.1
 
 ### Patch Changes